backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
sqlalchemy==2.0.25
asyncpg==0.29.0
redis==5.0.1
minio==7.2.3
neo4j==5.16.0
pymilvus==2.3.6
structlog==24.1.0
prometheus-client==0.19.0
simhash==2.1.2
datasketch==1.6.4
phonenumbers==8.13.29
python-multipart==0.0.9
tenacity==8.2.3
pytest==8.0.0
httpx==0.26.0
beautifulsoup4==4.12.3
newspaper3k==0.2.8
yt-dlp==2023.12.30
google-play-scraper==1.2.4
app-store-scraper==0.3.5
rapidfuzz==3.6.1
openai==1.12.0
sentence-transformers==2.3.1
umap-learn==0.5.5
hdbscan==0.8.33
bertopic==0.16.0
networkx==3.2.1
dowhy==0.11.1
spacy==3.7.2
presidio-analyzer==2.2.351
presidio-anonymizer==2.2.351
torch==2.1.2 --index-url https://download.pytorch.org/whl/cu118
transformers==4.37.2
faster-whisper==0.10.0
mlflow==2.10.0
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
pyyaml==6.0.1
